feat: 회원정보 수정 시작

+++ src/test/java/com/web/billim/order/domain/ProductOrderTest.java
@@ -19,7 +19,7 @@ class ProductOrderTest {
 			.build();
 
 		// when
-		int price = order.getPrice();
+		long price = order.getPrice();
 
 		// then
 		assertEquals(price, 6000);
